Scripts used to copy data from old location.
Note: these scripts are in the latest form, the location of the file
changes. I.e. this is not the full list of files copied over but a list
of files I so happened to have copied over latest.

new file:   comp_tetgenold.sh
new file:   comp_with_old.sh